Refactor and cleanup example2 app
