+++ login-workflow/src/new-architecture/contexts/RegistrationContext/provider.tsx
@@ -6,7 +6,7 @@
 import React, { useEffect } from 'react';
 import { RegistrationContextProviderProps } from './types';
 import { RegistrationContext } from './context';
-import { I18nextProvider, useTranslation } from 'react-i18next';
+import { I18nextProvider } from 'react-i18next';
 import { i18nRegistrationInstance } from './i18nRegistrationInstance';
 import Typography from '@mui/material/Typography';
 
@@ -14,7 +14,6 @@ export const RegistrationContextProvider: React.FC<React.PropsWithChildren<Regis
     props
 ) => {
     const { language, i18n = i18nRegistrationInstance } = props;
-    const { t } = useTranslation();
 
     useEffect(() => {
         void i18n.changeLanguage(language);
@@ -22,13 +21,7 @@ export const RegistrationContextProvider: React.FC<React.PropsWithChildren<Regis
 
     return (
         <I18nextProvider i18n={i18n}>
-            <RegistrationContext.Provider value={props}>
-                {/* @TODO: RIP out these next 3 lines once we have some stuff to test built out under the RegistrationContext */}
-                <Typography>{t('bluiRegistration:REGISTRATION.EULA.LOADING')}</Typography>
-                <Typography>{t('bluiRegistration:REGISTRATION.STEPS.COMPLETE')}</Typography>
-                <Typography>{t('bluiRegistration:test')}</Typography>
-                {props.children}
-            </RegistrationContext.Provider>
+            <RegistrationContext.Provider value={props}>{props.children}</RegistrationContext.Provider>
         </I18nextProvider>
     );
 };
